@@ -46,11 +46,11 @@
   "devDependencies": {
     "@types/chai": "^3.4.34",
     "@types/hammerjs": "^2.0.33",
-    "@types/jasmine": "2.5.41",
+    "@types/jasmine": "2.5.47",
     "@types/moment": "^2.13.0",
     "@types/node": "^7.0.11",
     "@types/protractor": "^4.0.0",
-    "@types/selenium-webdriver": "3.0.1",
+    "@types/selenium-webdriver": "3.0.3",
     "@types/source-map": "^0.5.0",
     "@types/uglify-js": "^2.0.27",
     "@types/webpack": "^2.2.12",
@@ -60,7 +60,7 @@
     "assets-webpack-plugin": "^3.4.0",
     "awesome-typescript-loader": "^3.1.2",
     "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "^0.27.3",
+    "css-loader": "^0.28.1",
     "del": "^2.2.2",
     "dotenv": "^4.0.0",
     "gulp": "^3.9.1",
@@ -89,10 +89,10 @@
     "ts-loader": "^2.0.3",
     "tslint": "^5.0.0",
     "tslint-loader": "^3.4.3",
-    "typescript": "2.2.2",
-    "webpack": "2.3.2",
+    "typescript": "2.3.2",
+    "webpack": "2.5.1",
     "webpack-dev-middleware": "^1.6.1",
-    "webpack-dev-server": "2.4.2",
+    "webpack-dev-server": "2.4.5",
     "webpack-md5-hash": "^0.0.5",
     "webpack-merge": "^4.1.0"
   }
